project_4/requirements.txt
gitpython>=3.1.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
nltk>=3.7
scikit-learn>=1.1.0
textblob>=0.17.0
click>=8.0.0
rich>=12.0.0
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0
gunicorn>=20.1.0